removed AS backends and instead rely on MultiJson for json decoding
