+++ activesupport/lib/active_support/json/decoding.rb
@@ -1,32 +1,31 @@
 require 'active_support/core_ext/module/attribute_accessors'
 require 'active_support/core_ext/module/delegation'
+require 'multi_json'
 
 module ActiveSupport
   # Look for and parse json strings that look like ISO 8601 times.
   mattr_accessor :parse_json_times
 
   module JSON
-    # Listed in order of preference.
-    DECODERS = %w(Yajl Yaml)
-
     class << self
-      attr_reader :parse_error
-      delegate :decode, :to => :backend
-
-      def backend
-        set_default_backend unless defined?(@backend)
-        @backend
-      end
-
-      def backend=(name)
-        if name.is_a?(Module)
-          @backend = name
+      def decode(json, options ={})
+        data = MultiJson.decode(json, options)
+        if ActiveSupport.parse_json_times
+          convert_dates_from(data)
         else
-          require "active_support/json/backends/#{name.to_s.downcase}"
-          @backend = ActiveSupport::JSON::Backends::const_get(name)
+          data
         end
-        @parse_error = @backend::ParseError
       end
+
+      def engine
+        MultiJson.engine
+      end
+      alias :backend :engine
+
+      def engine=(name)
+        MultiJson.engine = name
+      end
+      alias :backend= :engine=
 
       def with_backend(name)
         old_backend, self.backend = backend, name
@@ -35,15 +34,26 @@ module ActiveSupport
         self.backend = old_backend
       end
 
-      def set_default_backend
-        DECODERS.find do |name|
+      private
+
+      def convert_dates_from(data)
+        case data
+        when nil
+          nil
+        when DATE_REGEX
           begin
-            self.backend = name
-            true
-          rescue LoadError
-            # Try next decoder.
-            false
+            DateTime.parse(data)
+          rescue ArgumentError
+            data
           end
+        when Array
+          data.map! { |d| convert_dates_from(d) }
+        when Hash
+          data.each do |key, value|
+            data[key] = convert_dates_from(value)
+          end
+        else
+          data
         end
       end
     end
